+++ packages/website/src/templates/doc.jsx
@@ -1,22 +1,22 @@
-import { useLocation } from "@reach/router";
-import classNames from "classnames";
-import { graphql } from "gatsby";
-import SEO from "../components/SEO";
-import docsItems from "../data/docs";
-import guideItems from "../data/guides";
-import OctocatIcon from "../svg/octocat.svg";
-import slug from "../utils/slug";
-import Default from "../components/Default";
-import OpenSource from "../components/OpenSource";
-import React, { useEffect } from "react";
-import * as styles from "./doc.module.scss";
+import { useLocation } from '@reach/router';
+import classNames from 'classnames';
+import { graphql } from 'gatsby';
+import SEO from '../components/SEO';
+import docsItems from '../data/docs';
+import guideItems from '../data/guides';
+import OctocatIcon from '../svg/octocat.svg';
+import slug from '../utils/slug';
+import Default from '../components/Default';
+import OpenSource from '../components/OpenSource';
+import React, { useEffect } from 'react';
+import * as styles from './doc.module.scss';
 
 const getPagination = (permalink, items) => {
   const flattenedItems = [];
 
   const flatItems = (i) => {
     i.forEach((child) => {
-      if (child.url.includes("#")) return;
+      if (child.url.includes('#')) return;
       flattenedItems.push(child);
       if (child.children) {
         flatItems(child.children);
@@ -55,19 +55,19 @@ const SidebarItem = ({ item }) => {
   );
 };
 
-const Doc = ({ data, location, children }) => {
+const Doc = ({ data, children }) => {
   const { title } = data.mdx.frontmatter;
   const { permalink, pageType } = data.mdx.fields;
-  const category = pageType === "guides" ? "Guides" : "Documentation";
+  const category = pageType === 'guides' ? 'Guides' : 'Documentation';
 
-  const items = pageType === "guides" ? guideItems : docsItems;
+  const items = pageType === 'guides' ? guideItems : docsItems;
   const { prev, next } = getPagination(permalink, items);
 
   useEffect(() => {
     docsearch({
-      apiKey: "281facf513620e95600126795a00ab6c",
-      indexName: "quilljs",
-      inputSelector: ".search-item input",
+      apiKey: '281facf513620e95600126795a00ab6c',
+      indexName: 'quilljs',
+      inputSelector: '.search-item input',
       debug: false,
     });
   }, []);
@@ -88,14 +88,14 @@ const Doc = ({ data, location, children }) => {
             </ul>
           </div>
           <div id="docs-container" className="nine columns">
-            <div className={classNames("row", styles.breadcrumbRow)}>
+            <div className={classNames('row', styles.breadcrumbRow)}>
               <div className={styles.breadcrumb}>
                 <span>{category}:</span>
                 <span>{title}</span>
               </div>
               <a
                 className={styles.editLink}
-                href={`${data.site.siteMetadata.github}${location.pathname}`}
+                href={data.mdx.fields.githubPath}
                 target="_blank"
                 title="Edit on GitHub"
               >
@@ -104,10 +104,10 @@ const Doc = ({ data, location, children }) => {
               </a>
             </div>
             <hr />
-            <div id="content-container">
+            <article id="content-container" className={styles.content}>
               <h1 id={slug(title)}>{title}</h1>
               {children}
-            </div>
+            </article>
             <div className="row" id="pagination-container">
               {prev && (
                 <a className="prev" href={prev.url}>
@@ -143,16 +143,12 @@ const Doc = ({ data, location, children }) => {
 
 export const query = graphql`
   query ($id: String) {
-    site {
-      siteMetadata {
-        github
-      }
-    }
     mdx(id: { eq: $id }) {
       fields {
         slug
         permalink
         pageType
+        githubPath
       }
       frontmatter {
         title